Important info (Read Description)
For this particular project and due to its complexity, please
Skype/text me BEFORE you add anything.

Reading the commit history of the game to see what has changed or not
is not enough, because it does not explain the purpose on WHY certain
things are added.  The design for this particular strategy is complex,
it is not straight forward like an action game, and therefore ambiguity
is not affordable.

GameManager is a script that I added in the scripts folder, which I
attached to an empty gameObject called gameManager in scene1. What I
just typed is an example of how a commit should be described.  Yes it
should be this detailed and it is NOT hard.  I don't want an argument
to spark for how commits and keeping each other updated should be done.
 There is NO JOKING around for this one!  We cannot afford to be lazy
or make it "easy" on communicate, because ambiguity for an ambitious
project like this is extremely dangerous.  For the case of development,
since the entire team can fall behind if one person doesn't understand
the information correctly.  For other things I agree it is based on
vote.  But for this case everyone has to be happy, not just the
majority.  Just please do this for the sake of the project not falling
behind.  When I have a question, I expect to be helped.  The "i don't
need to hand-hold you" is not an excuse, because everyone is expected
to hand-hold another, and handhold.  All of us have way too many
classes on our hand that we cannot afford an exact simulation of
working on a project in a company, because the people there have only
the project to focus on, not on general ed classes like history and
humanities and essay composition.  So please, just allow this only
accommodation so that no one in the team has to suffer from ambiguity
in this project.