Lots of issues
I already finished the basic dragging and dropping as well as creating
the terrain and drag and drop item to demonstrate it works.  But, one
of you has an older version of unity not compatible with my version.
Check the repo to make sure all of you have the same version as me.
All my hard work got deleted so I made a second folder containing all
my work just for the setup screen.  We need to meet in person to carry
my work over to the main unity project, update the version, and
configure git to ignore certain files.