PieceStruct array in PlayerSetup in more progress
NOTE:  Each Unit needs a constructor.  the PieceStruct object consists
of a Unit object and a GameObject prefab object,  for optimization.
However, the constructor for each unit subclass RE-INSTANTIATES.  The
re-instantiation is ONLY for during the PlayerSetup scene.  Do not
confuse the two variables together, and be CAREFUL whenever you respawn
a piece so the values actually reset, upgrade, etc!!!!